requirements.txt
flask==2.3.3
mlflow==2.8.1
tensorflow==2.12
pillow==10.1.0